+++ certbot/main.py
@@ -15,7 +15,7 @@ from acme.magic_typing import Union  # pylint: disable=unused-import, no-name-in
 
 import certbot
 from certbot._internal import account
-from certbot import cert_manager
+from certbot._internal import cert_manager
 from certbot import cli
 from certbot import client
 from certbot import configuration
