--- docs/api/cert_manager.rst
@@ -1,5 +1,0 @@
-:mod:`certbot.cert_manager`
--------------------------------
-
-.. automodule:: certbot.cert_manager
-   :members:
